packages/sensor/pyproject.toml
[tool.poetry]
name = "sensor"
version = "0.1.0"
description = "Temperature and humidity sensor abstraction"
authors = []
license = "GPL-3.0-only"
readme = "README.md"

[[tool.poetry.packages]]
include = "sensor"

[tool.poetry.scripts]
sensor = "sensor.cli.main:cli"

[tool.poetry.dependencies]
python = ">=3.9"
click = "^8.1.7"
grpcio-tools = "^1.60.0"
pytest-asyncio = "^0.23.2"

[tool.poetry.group.dev.dependencies]
ruff = "0.1.8"
pytest = "7.4.3"
black = "^23.12.0"
mypy = "^1.7.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
extend-exclude = ["**/*_pb2*"]

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "I"]

[tool.black]
extend-exclude = ".*_pb2*"

[tool.mypy]
exclude = [".*_pb2*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
